backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
pyyaml==6.0.2
orjson==3.10.7







